# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kulturbytes Research Planner",
    "version": "0.1.0"
  },
  "paths": {
...
    "/plan": {
      "post": {
        "summary": "Plan",
        "operationId": "plan_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PlannedResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ClarificationResponse"
                    }
                  ],
                  "title": "Response Plan Plan Post",
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "plan": "#/components/schemas/PlannedResponse",
                      "needs_clarification": "#/components/schemas/ClarificationResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v5/plan": {
      "post": {
        "summary": "Analytical Plan",
        "operationId": "analytical_plan_v5_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AnalyticalResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnalyticalClarification"
                    }
                  ],
                  "title": "Response Analytical Plan V5 Plan Post",
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "plan": "#/components/schemas/AnalyticalResponse",
                      "needs_clarification": "#/components/schemas/AnalyticalClarification"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v6/plan": {
      "post": {
        "summary": "Geographic Plan",
        "operationId": "geographic_plan_v6_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/GeographicResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GeographicClarification"
                    }
                  ],
                  "title": "Response Geographic Plan V6 Plan Post",
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "plan": "#/components/schemas/GeographicResponse",
                      "needs_clarification": "#/components/schemas/GeographicClarification"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v7/plan": {
      "post": {
        "summary": "Research Plan V7",
        "operationId": "research_plan_v7_v7_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseV7"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v4/plan": {
      "post": {
        "summary": "Domain Plan",
        "operationId": "domain_plan_v4_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEnvelopeV4"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalyticalResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "research-query-plan-v5",
            "title": "Schema Version"
          },
          "prompt_version": {
            "type": "string",
            "const": "research-planner-v10",
            "title": "Prompt Version"
          },
          "model": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Model"
          },
          "plan": {
            "$ref": "#/components/schemas/AnalyticalQueryPlan"
          },
          "reference_date": {
            "type": "string",
            "format": "date",
            "title": "Reference Date"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "title": "Timezone"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/AnalyticalDiagnostics"
          },
          "kind": {
            "type": "string",
            "const": "plan",
            "title": "Kind",
            "default": "plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "prompt_version",
          "model",
          "plan",
          "reference_date",
          "timezone",
          "diagnostics"
        ],
        "title": "AnalyticalResponse"
      },
...
      "ClarificationResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "research-query-plan-v3",
            "title": "Schema Version",
            "default": "research-query-plan-v3"
          },
          "prompt_version": {
            "type": "string",
            "const": "research-planner-v7",
            "title": "Prompt Version"
          },
          "model": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Model"
          },
          "plan": {
            "$ref": "#/components/schemas/ResearchQueryPlan"
          },
          "reference_date": {
            "type": "string",
            "format": "date",
            "title": "Reference Date"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "title": "Timezone"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/PlanDiagnostics"
          },
          "kind": {
            "type": "string",
            "const": "needs_clarification",
            "title": "Kind",
            "default": "needs_clarification"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prompt_version",
          "model",
          "plan",
          "reference_date",
          "timezone",
          "diagnostics"
        ],
        "title": "ClarificationResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "PlanRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Timezone",
            "default": "Europe/Berlin"
          },
          "language": {
            "type": "string",
            "enum": [
              "de",
              "da",
              "en",
              "auto"
            ],
            "title": "Language",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "PlanRequest"
      },
...
      "PlannedResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "research-query-plan-v3",
            "title": "Schema Version",
            "default": "research-query-plan-v3"
          },
          "prompt_version": {
            "type": "string",
            "const": "research-planner-v7",
            "title": "Prompt Version"
          },
          "model": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Model"
          },
          "plan": {
            "$ref": "#/components/schemas/ResearchQueryPlan"
          },
          "reference_date": {
            "type": "string",
            "format": "date",
            "title": "Reference Date"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "title": "Timezone"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/PlanDiagnostics"
          },
          "kind": {
            "type": "string",
            "const": "plan",
            "title": "Kind",
            "default": "plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prompt_version",
          "model",
          "plan",
          "reference_date",
          "timezone",
          "diagnostics"
        ],
        "title": "PlannedResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Internal service key, not a user token",
        "scheme": "bearer"
      }
    }
  }
}